server/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Real-time Multiplayer Chess Platform"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    # Core Django
    "Django==5.0.1",
    "djangorestframework==3.14.0",

    # Media
    "pillow==12.0.0",

    # JWT Authentication
    "djangorestframework-simplejwt==5.3.1",

    # Google OAuth
    "google-auth==2.23.0",
    "google-auth-oauthlib==1.1.0",
    "google-auth-httplib2==0.1.1",

    # Redis / Cache / PubSub
    "redis[hiredis]==5.0.1",
    "django-redis==5.4.0",

    # Actor Model (Replaces Celery)
    "thespian>=3.10.1",

    # Database
    "psycopg2-binary==2.9.9",
    "dj-database-url==2.1.0",

    # WebSockets & Async
    "channels==4.0.0",
    "channels-redis==4.1.0",
    "daphne==4.0.0",

    # CORS
    "django-cors-headers==4.3.1",

    # Environment
    "python-dotenv==1.0.0",
]